requirements.txt

fastapi==0.112.1
uvicorn==0.30.6
motor==3.5.1
pydantic==2.8.2
pydantic_core==2.20.1
pymongo==4.8.0
python-jose==3.3.0